try to get migrations to run

+++ db/migrate/20141029173233_add_missing_unique_indices.acts_as_taggable_on_engine.rb
@@ -1,7 +1,7 @@
 # This migration comes from acts_as_taggable_on_engine (originally 2)
 class AddMissingUniqueIndices < ActiveRecord::Migration
   def self.up
-    remove_index :tags, :name
+    remove_index :tags, :name if index_exists?(:tags, :name)
     add_index :tags, :name, unique: true
 
     remove_index :taggings, :tag_id
